Modularize code to compute Spearman correlations
